Add files via upload
In this project, we analyzed data from the Chicago Transit Authority (CTA) to understand how ridership patterns have changed over the years. Our goal was to find ways to make CTA transportation more efficient. We used Pyspark and Python to process the data, creating statistical graphs to visualize trends. The source data was stored in an S3 bucket, and we employed Python for data analysis and scripting.